Refactor interactive ports
