+++ pyseq2/com/async_com.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import asyncio
-import logging
 import time
 from asyncio import Future, StreamReader, StreamWriter
 from dataclasses import dataclass
@@ -263,28 +262,3 @@ class COM:
 
     async def wait(self) -> None:
         return await self._read_queue.join()
-
-
-async def interactive():
-    import aioconsole
-    from pyseq2.utils.ports import get_ports
-    from rich.logging import RichHandler
-
-    logging.basicConfig(
-        level="NOTSET",
-        format="[yellow]%(name)-10s[/] %(message)s",
-        datefmt="[%X]",
-        handlers=[RichHandler(rich_tracebacks=True, markup=True)],
-    )
-
-    name: SerialInstruments = await aioconsole.ainput("Instrument? ")
-    com = await COM.ainit(name, getattr(get_ports(), name))
-    while True:
-        await asyncio.sleep(0.2)
-        line = await aioconsole.ainput("Command? ")
-        await aioconsole.aprint()
-        await com.send(line)
-
-
-if __name__ == "__main__":
-    asyncio.run(interactive())
